frontend feat: listing when added
